Rename Microsoft.AspNet.TelemetryCorrelation -> OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule. Added project to the solution + updated references & namespaces. (#2238)
