+++ examples/AspNet/Web.config
@@ -23,7 +23,7 @@
 			<add name="ExtensionlessUrlHandler-Integrated-4.0" path="*." verb="*" type="System.Web.Handlers.TransferRequestHandler" preCondition="integratedMode,runtimeVersionv4.0"/>
 		</handlers>
 		<modules>
-			<add name="TelemetryCorrelationHttpModule" type="Microsoft.AspNet.TelemetryCorrelation.TelemetryCorrelationHttpModule, Microsoft.AspNet.TelemetryCorrelation" preCondition="integratedMode,managedHandler"/>
+			<add name="TelemetryCorrelationHttpModule" type="OpenTelemetry.Instrumentation.AspNet.TelemetryCorrelationHttpModule, OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule" preCondition="integratedMode,managedHandler"/>
 		</modules>
 	</system.webServer>
 	<runtime>
