+++ src/OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule/web.config.install.xdt
