+++ test/OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule.Tests/ActivityExtensionsTest.cs
@@ -14,7 +14,7 @@
 // limitations under the License.
 // </copyright>
 
-namespace Microsoft.AspNet.TelemetryCorrelation.Tests
+namespace OpenTelemetry.Instrumentation.AspNet.Tests
 {
     using System.Collections.Generic;
     using System.Collections.Specialized;
