+++ test/OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule.Tests/PropertyExtensions.cs
@@ -1,4 +1,4 @@
-// <copyright file="HttpParseResult.cs" company="OpenTelemetry Authors">
+// <copyright file="PropertyExtensions.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,24 +14,15 @@
 // limitations under the License.
 // </copyright>
 
-namespace Microsoft.AspNet.TelemetryCorrelation
+namespace OpenTelemetry.Instrumentation.AspNet.Tests
 {
-    // Adoptation of code from https://github.com/aspnet/HttpAbstractions/blob/07d115400e4f8c7a66ba239f230805f03a14ee3d/src/Microsoft.Net.Http.Headers/HttpParseResult.cs
-    internal enum HttpParseResult
+    using System.Reflection;
+
+    internal static class PropertyExtensions
     {
-        /// <summary>
-        /// Parsed successfully.
-        /// </summary>
-        Parsed,
-
-        /// <summary>
-        /// Was not parsed.
-        /// </summary>
-        NotParsed,
-
-        /// <summary>
-        /// Invalid format.
-        /// </summary>
-        InvalidFormat,
+        public static object GetProperty(this object obj, string propertyName)
+        {
+            return obj.GetType().GetTypeInfo().GetDeclaredProperty(propertyName)?.GetValue(obj);
+        }
     }
 }
